requirements.txt
# Core dependencies
feedparser==6.0.11
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Scheduling
schedule==1.2.1

# Translation (if using translator.py with MyMemory API)
urllib3==2.2.2

# Time and encoding support
python-dateutil==2.9.0.post0
